+++ package.json
@@ -58,6 +58,7 @@
     "mobx-react": "^7.2.1",
     "mobx-react-form": "^3.2.0",
     "next": "^13.0.3",
+    "next-absolute-url": "^1.2.2",
     "object-hash": "^2.2.0",
     "plaiceholder": "^2.5.0",
     "pure-react-carousel": "^1.27.8",
